+++ docs/version-history.md
@@ -8,6 +8,7 @@ Version 0.3 remains in beta testing. The most significant change from version 0.
 
   * Module **props** renamed as **properties**;
   * Module **coeffs** renamed as **coefficients**;
+  * Class **CoeffLib** renamed as **CoefficientLibrary**;
   * Added docstrings to all functions in **coefficients** module;
   * Added **matrix** module with alternative matrix-based model implementation;
   * Added **none** function to **jfuncs** module to ignore unsymmetrical mixing.